refactor(html): fold the head-tag builders into the html syntax module
They only build strings out of escapeAttribute/escapeText, so they belong
next to those rather than in a file of their own. Both callers already
reach the syntax module, so this drops a require instead of adding one.
