+++ lib/html/HtmlModulesPlugin.js
@@ -39,7 +39,6 @@ const {
 const removeBOM = require("../util/removeBOM");
 const HtmlGenerator = require("./HtmlGenerator");
 const HtmlModule = require("./HtmlModule");
-const { buildHeadTags } = require("./headTags");
 
 const getMimeTypes = memoize(() => require("../util/mimeTypes"));
 
@@ -437,7 +436,7 @@ class HtmlModulesPlugin {
 						scripts.push(`<script${scriptAttr} src="${r}"></script>`);
 					}
 				}
-				const headTags = buildHeadTags(htmlObj);
+				const headTags = getHtmlSyntax().buildHeadTags(htmlObj);
 				// Injected into webpack-generated pages only; each icon's href is a
 				// normal request the parser turns into a hashed asset.
 				const faviconTag = resolveFaviconLinks(htmlObj.favicon, name)
