+++ lib/html/HtmlParser.js
@@ -25,17 +25,19 @@ const {
 	webpackCommentRegExp
 } = require("../util/magicComment");
 const memoize = require("../util/memoize");
-const { baseTag, buildHeadTags, metaTag } = require("./headTags");
 const {
 	NS_HTML,
 	NS_SVG,
 	NodeType,
 	SVG_TAG_ADJUST,
 	SourceProcessor,
+	baseTag,
+	buildHeadTags,
 	decodeEntities,
 	escapeAttribute,
 	escapeText,
 	isAsciiWhitespace,
+	metaTag,
 	parseCssUrls,
 	parseMsapplicationTask,
 	parseSrc,
