+++ lib/html/syntax.js
@@ -8,6 +8,7 @@
 const GenericSourceProcessor = require("../util/SourceProcessor");
 
 /** @typedef {import("../css/syntax").CssEnvironment} CssEnvironment */
+/** @typedef {import("../../declarations/WebpackOptions").OutputHtmlOptions} OutputHtmlOptions */
 const {
 	ADDRESS_DIV_P,
 	APPLET_MARQUEE_OBJECT,
@@ -4612,6 +4613,58 @@ const escapeText = (s) => {
 		last = i + 1;
 	}
 	return out + s.slice(last);
+};
+
+/**
+ * @param {string} name meta name
+ * @param {string} content meta content
+ * @returns {string} meta tag
+ */
+const metaTag = (name, content) => {
+	// og: uses property=; all others (including twitter:) use name=
+	const attr = name.startsWith("og:")
+		? `property="${escapeAttribute(name)}"`
+		: `name="${escapeAttribute(name)}"`;
+	return `<meta ${attr} content="${escapeAttribute(content)}">`;
+};
+
+/**
+ * @param {string | { href: string, target?: string }} base base option
+ * @returns {string} base tag
+ */
+const baseTag = (base) => {
+	const href = typeof base === "string" ? base : base.href;
+	const targetAttr =
+		typeof base === "object" && base.target
+			? ` target="${escapeAttribute(base.target)}"`
+			: "";
+	return `<base href="${escapeAttribute(href)}"${targetAttr}>`;
+};
+
+/**
+ * Serializes the head tags `output.html` asks for, in spec order: charset,
+ * base, meta, title. Only for pages built from scratch — an authored page
+ * merges the same options during `parse`, where its own tags win.
+ * @param {OutputHtmlOptions} opts html options
+ * @returns {string} head tags string
+ */
+const buildHeadTags = (opts) => {
+	let out = "";
+	const meta = opts.meta;
+	if (meta && meta.charset) {
+		out += `<meta charset="${escapeAttribute(meta.charset)}">`;
+	}
+	if (opts.base) out += baseTag(opts.base);
+	if (meta) {
+		for (const [name, content] of Object.entries(meta)) {
+			if (name === "charset") continue;
+			out += metaTag(name, content);
+		}
+	}
+	if (opts.title) {
+		out += `<title>${escapeText(opts.title)}</title>`;
+	}
+	return out;
 };
 
 // cspell:ignore definitionurl malignmark mglyph selectedcontent megamorphic attributeless rowspan imagesizes novalidate maxlength
@@ -10041,12 +10094,15 @@ module.exports.QUOTE_SINGLE = QUOTE_SINGLE;
 // the adjusted camelCase names the AST carries for foreign content.
 module.exports.SVG_TAG_ADJUST = SVG_TAG_ADJUST;
 module.exports.SourceProcessor = SourceProcessor;
+module.exports.baseTag = baseTag;
+module.exports.buildHeadTags = buildHeadTags;
 module.exports.decodeEntities = decodeEntities;
 module.exports.escapeAttribute = escapeAttribute;
 module.exports.escapeText = escapeText;
+module.exports.isAsciiWhitespace = isSpace;
+module.exports.metaTag = metaTag;
 // WHATWG "ASCII whitespace" (tab / LF / FF / CR / space) — the tokenizer's
 // whitespace class, exported under the spec's name.
-module.exports.isAsciiWhitespace = isSpace;
 module.exports.parseCssUrls = parseCssUrls;
 module.exports.parseHtml = parseHtml;
 module.exports.parseMsapplicationTask = parseMsapplicationTask;
